split slow group tests
